add simple input error handling for dbUtils.clearCollection()

+++ utils/dbUtils.js
@@ -8,6 +8,9 @@ const dbUtils = {
   },
 
   async clearCollection(modelName) {
+    if (typeof modelName !== 'string') {
+      throw new TypeError('First parameter should be of string data type')
+    }
     const ModelClass = this.getModelClass(modelName)
 
     await ModelClass.deleteMany({})
